requirements.txt
loguru
openai
pydantic
pytest
python-dotenv
requests
httpx
black
SQLAlchemy
beautifulsoup4
pandas
transformers
streamlit
psycopg2
langchain_community
tiktoken
python-telegram-bot
apscheduler
